+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 0.3a0
+* Added zero method
+* Changed `RPiStepper` class to `Motor` class
+* Added RPistepper shell
+* in progress
+
 ## 0.2a0
 * Fixed documentation
 * GPIO output callings now in list/tuple format
@@ -7,3 +13,6 @@
 
 ## 0.1a0
 * Initial release
+
+# Issues
+* Piping `waitkey` in the shell doesn't stops the excecution